Move shared config method to module root
